backend/requirements.txt
### Required dependencies ###

django~=4.2
psycopg[binary]~=3.1
djangorestframework~=3.14
drf-spectacular~=0.26
drf-yasg==1.21.7
mysqlclient


# Filtering
django-filter~=23.3

# Cors headers
django-cors-headers~=4.3

# Linter
flake8

